+++ CHANGES.md
@@ -18,6 +18,7 @@
 - Fix problem with join nodes [#1422](https://github.com/penpot/penpot/issues/1422)
 - After team onboarding importing a file will import into the team drafts [Taiga #2408](https://tree.taiga.io/project/penpot/issue/2408)
 - Fix problem exporting shapes from handoff mode [Taiga #2386](https://tree.taiga.io/project/penpot/issue/2386)
+- Fix lock/hide elements in context menu when multiples shapes selected [Taiga #2340](https://tree.taiga.io/project/penpot/issue/2340)
 
 ### :arrow_up: Deps updates
 ### :heart: Community contributions by (Thank you!)